Add devices argument to RRef constructor
PR #54932 fixed the CUDA RPC for RRef when RRef is created through
RPC. But besides that use case, RRef can also be created locally
by directly passing in a value, which would bypass the CUDA stream
synchronization in #54932.

This commit covers the above gap by adding a `devices` argument
to RRef constructor. The RRef will then use this argument to
choose between `CUDAFutre` and `ivalue::Future` to hold the value.
When `devices` is specified and non-empty, `CUDAFuture` will be
used, and the `devices` will be passed to that `CUDAFuture`.

ghstack-source-id: 0dc00c79a8c28d3d53eba8b1ecac0fba04e8b0d4
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57085