Simplify CUDAMultiStreamGuard
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57048

CUDAMultiStreamGuard had a default constructor and a `original_devices()` method which were only used in a test. I'm removing them here to simplify the API and make it easier to manipulate this class later. One extra benefit is that this class used to get and store the current stream of _all_ devices, whereas now it only does so for the relevant devices.
ghstack-source-id: 127713136

Differential Revision: [D28029160](https://our.internmc.facebook.com/intern/diff/D28029160/)