[JIT] Add torch._C.ScriptDict
**Summary**
This commit adds `torch._C.ScriptDict`, a dictionary type that has reference
semantics across the Python/TorchScript boundary. That is, modifications
made to instances of `torch._C.ScriptDict` in TorchScript are visible in
Python even when it is not returned from the function. Instances can be
constructed by passing an instance of a Python dictionary to
`torch.jit.script`. In the case of an empty dictionary, its type is
assumed to be `Dict[str, Tensor]` to be consistent with the handling of
empty dictionaries in TorchScript source code.

`torch.jit.dict` is implemented using a modified version of pybind's `stl_bind.h`-style bindings attached to `ScriptDict`, `ScriptDictIterator` and `ScriptDictKeyIterator`, wrapper classes around `c10::impl::GenericDict` and `c10::impl::GenericDict::iterator`. These bindings allow instances of `torch._C.ScriptDict` to be used as if it were a regular `dict` Python. Reference semantics are achieved by simply retrieving the `IValue` contained in `ScriptDict` in `toIValue` (invoked when converting Python arguments to `IValues` before calling TorchScript code).

**Test Plan**
This commit adds `TestScriptDict` to `test_list_dict.py`, a set of tests
that check that all of the common dictionary operations are supported
and that instances have reference semantics across the
Python/TorchScript boundary.

The performance benefits have also been tested vs. trunk on a simple
benchmark that creates a large dict and adds an element to it in
TorchScript:
```
@torch.jit.script
def dict_append(a: Dict[int, int], k: int, v: int):
    a[k] = v
    return a

    d = {i: i for i in range(1000000)}
    t1 = time.monotonic()
    d = dict_append(d, -1, -1)
    t2 = time.monotonic()
    print(t2-t1)
```
This involves two copies (Python -> TS -> Python) on trunk and only one
with this commit. This commit speeds up this sample by 75%.

ghstack-source-id: 640b6fed123783ca282212bb2f0befab8432580d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52659