[TensorExpr] Remove `dtype_` and add `buf_` fields to `CodeGen::BufferArg`.
`BufferArg` is used to describe parameters passed to the codegen: it
indicates whether the parameter is a var or a buf and holds a pointer to
the corresponding var/buf. Both var and buf contain dtype, and thus
duplicating it in BufferArg is unnecessary - we can always get it from
the var/buf. Hence we're removing dtype_ field from BufferArg in this
PR. We're also adding a `buf_` field here: this is done so that
BufferArg truly has all the info about the parameter.

ghstack-source-id: e9e9ff211b1708ef2ef1dfe1c191964ba0e9c18d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57382