[TensorExpr] Add methods for inspecting generated code in `TensorExprKernel`.
The new methods allow to peak into bufferArgs which describe parameters
that codegen expects. This description includes info whether a given
parameter is a scalar var or a buffer and in case it's a buffer allows
to get the corresponding `Buf*` pointer from which we could get the
expected sizes.

ghstack-source-id: 76d7c4534293a2a9826285837bb3842777e38512
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57074