[quant][graphmode][fx] Skip observering boolean Tensors
Summary:
Skip observing the input for masked_fill. Currently we don't have a way to
query the type of Proxy in GraphModule, hopefully we should have the functionality to annotate the type,
we'll need to annotate a Proxy to be a boolean Tensor to remove this hack.

Test Plan:
python test/test_quantization.py TestQuantizeFxOps.test_boolean_tensor

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: dbe46f6dc9a3a7db9daaa1a1c75843c548240008
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57375