[RPC Framework] Create a separate remote module template when moving CPU tensors to a cuda device is not enabled
Pull Request resolved: https://github.com/pytorch/pytorch/pull/57413

An internal test fails because somehow `Tuple[()]` is not considered compatible with `Tuple[Any]` in TorchScript, even if the code that involves this type of variables is not executed at all.

Therefore, create separate templates for instantiation to avoid typing check failure. This can address the FIXME left in https://github.com/pytorch/pytorch/pull/57288

#Closes: https://github.com/pytorch/pytorch/issues/51670
ghstack-source-id: 127917180

Differential Revision: [D28138864](https://our.internmc.facebook.com/intern/diff/D28138864/)